Two-sample Test for High-Dimensional Means by Cai, Liu, and Xia (2014)

Given two multivariate data \(X\) and \(Y\) of same dimension, it tests $$H_0 : \mu_x = \mu_y\quad vs\quad H_1 : \mu_x \neq \mu_y$$ using the procedure by Cai, Liu, and Xia (2014). Since the test is equivalent to testing $$H_0 : \Omega(\mu_x - \mu_y)=0$$, if \(\Omega\), an inverse covariance (or precision), is not provided, it finds a sparse precision matrix via CLIME estimator. Also, if two samples are assumed to have different covariance structure, it uses weighting scheme for adjustment.

mean2.2014CLX(X, Y, Omega = NULL, cov.equal = TRUE)

Arguments

X
an \((n_x \times p)\) data matrix of 1st sample.

Y
an \((n_y \times p)\) data matrix of 2nd sample.

Omega
precision matrix; if NULL, it applies CLIME estimation. Otherwise, a \((p\times p)\) inverse covariance should be provided.

cov.equal
a logical to determine homogeneous covariance assumption.

Value

a (list) object of S3 class htest containing:

statistic

a test statistic.

p.value

\(p\)-value under \(H_0\).

alternative

alternative hypothesis.

method

name of the test.

data.name

name(s) of provided sample data.

References

Cai TT, Liu W, Xia Y (2014). “Two-sample test of high dimensional means under dependence.” Journal of the Royal Statistical Society: Series B (Statistical Methodology), 76(2), 349--372. ISSN 13697412.

Examples

## CRAN-purpose small example
smallX = matrix(rnorm(10*3),ncol=3)
smallY = matrix(rnorm(10*3),ncol=3)
mean2.2014CLX(smallX, smallY) # run the test
#> 
#> 	Two-sample Mean Test with Equal Precisions with CLIME estimate by Cai,
#> 	Liu, and Xia (2014).
#> 
#> data:  smallX and smallY
#> statistic = 1.984, p-value = 0.4505
#> alternative hypothesis: two means are different.
#> 

if (FALSE) {
## empirical Type 1 error 
niter   = 100
counter = rep(0,niter)  # record p-values
for (i in 1:niter){
  X = matrix(rnorm(50*5), ncol=10)
  Y = matrix(rnorm(50*5), ncol=10)
  
  counter[i] = ifelse(mean2.2014CLX(X, Y)$p.value < 0.05, 1, 0)
}

## print the result
cat(paste("\n* Example for 'mean2.2014CLX'\n","*\n",
"* number of rejections   : ", sum(counter),"\n",
"* total number of trials : ", niter,"\n",
"* empirical Type 1 error : ",round(sum(counter/niter),5),"\n",sep=""))
}