One-sample Simultaneous Likelihood Ratio Test of Mean and Variance

Given two univariate samples \(x\) and \(y\), it tests $$H_0 : \mu_x = \mu_0, \sigma_x^2 = \sigma_0^2 \quad vs \quad H_1 : \textrm{ not } H_0$$ using likelihood ratio test.

mvar1.LRT(x, mu0 = 0, var0 = 1)

Arguments

x

a length-\(n\) data vector.

mu0

hypothesized mean \(\mu_0\).

var0

hypothesized variance \(\sigma_0^2\).

Value

a (list) object of S3 class htest containing:

statistic

a test statistic.

p.value

\(p\)-value under \(H_0\).

alternative

alternative hypothesis.

method

name of the test.

data.name

name(s) of provided sample data.

Examples

## CRAN-purpose small example
mvar1.LRT(rnorm(10))
#> 
#> 	One-sample Simultaneous Likelihood Ratio Test of Mean and Variance.
#> 
#> data:  rnorm(10)
#> statistic = 0.01297, p-value = 0.9935
#> alternative hypothesis: true mean and variance of x are different from mu0 and var0.
#> 

if (FALSE) {
## empirical Type 1 error 
niter   = 1000
counter = rep(0,niter)  # record p-values
for (i in 1:niter){
  x = rnorm(100)  # sample x from N(0,1)
  
  counter[i] = ifelse(mvar1.LRT(x)$p.value < 0.05, 1, 0)
}

## print the result
cat(paste("\n* Example for 'mvar1.LRT'\n","*\n",
"* number of rejections   : ", sum(counter),"\n",
"* total number of trials : ", niter,"\n",
"* empirical Type 1 error : ",round(sum(counter/niter),5),"\n",sep=""))
}