Robust Jarque-Bera Test of Univariate Normality by Gel and Gastwirth (2008)

Given an univariate sample \(x\), it tests $$H_0 : x\textrm{ is from normal distribution} \quad vs\quad H_1 : \textrm{ not } H_0$$ using a test procedure by Gel and Gastwirth (2008), which is a robustified version Jarque-Bera test.

norm.2008RJB(x, C1 = 6, C2 = 24, method = c("asymptotic", "MC"), nreps = 2000)

Arguments

x
a length-\(n\) data vector.

C1
a control constant. Authors proposed \(C1=6\) for nominal level of \(\alpha=0.05\).

C2
a control constant. Authors proposed \(C2=24\) for nominal level of \(\alpha=0.05\).

method
method to compute \(p\)-value. Using initials is possible, "a" for asymptotic for example.

nreps
the number of Monte Carlo simulations to be run when method="MC".

Value

a (list) object of S3 class htest containing:

statistic

a test statistic.

p.value

\(p\)-value under \(H_0\).

alternative

alternative hypothesis.

method

name of the test.

data.name

name(s) of provided sample data.

References

Gel YR, Gastwirth JL (2008). “A robust modification of the Jarque–Bera test of normality.” Economics Letters, 99(1), 30--32. ISSN 01651765.

Examples

## generate samples from uniform distribution
x = runif(28)

## test with both methods of attaining p-values
test1 = norm.2008RJB(x, method="a") # Asymptotics
test2 = norm.2008RJB(x, method="m") # Monte Carlo