One-sample Simultaneous Test of Mean and Covariance by Liu et al. (2017)

Given a multivariate sample \(X\), hypothesized mean \(\mu_0\) and covariance \(\Sigma_0\), it tests $$H_0 : \mu_x = \mu_0 \textrm{ and } \Sigma_x = \Sigma_0 \quad vs\quad H_1 : \textrm{ not } H_0$$ using the procedure by Liu et al. (2017).

sim1.2017Liu(X, mu0 = rep(0, ncol(X)), Sigma0 = diag(ncol(X)))

Arguments

X

an \((n\times p)\) data matrix where each row is an observation.

mu0

a length-\(p\) mean vector of interest.

Sigma0

a \((p\times p)\) given covariance matrix.

Value

a (list) object of S3 class htest containing:

statistic

a test statistic.

p.value

\(p\)-value under \(H_0\).

alternative

alternative hypothesis.

method

name of the test.

data.name

name(s) of provided sample data.

References

Liu Z, Liu B, Zheng S, Shi N (2017). “Simultaneous testing of mean vector and covariance matrix for high-dimensional data.” Journal of Statistical Planning and Inference, 188, 82--93. ISSN 03783758.

Examples

## CRAN-purpose small example
smallX = matrix(rnorm(10*3),ncol=3)
sim1.2017Liu(smallX) # run the test
#> 
#> 	One-sample Simultaneous Test of Mean and Covariance by Liu et al.
#> 	(2017)
#> 
#> data:  smallX
#> statistic = -0.67018, p-value = 0.5027
#> alternative hypothesis: both mean and covariance are not equal to mu0 and Sigma0.
#> 

if (FALSE) {
## empirical Type 1 error 
niter   = 1000
counter = rep(0,niter)  # record p-values
for (i in 1:niter){
  X = matrix(rnorm(50*10), ncol=10)
  counter[i] = ifelse(sim1.2017Liu(X)$p.value < 0.05, 1, 0)
}

## print the result
cat(paste("\n* Example for 'sim1.2017Liu'\n","*\n",
"* number of rejections   : ", sum(counter),"\n",
"* total number of trials : ", niter,"\n",
"* empirical Type 1 error : ",round(sum(counter/niter),5),"\n",sep=""))
}